simplify `Writer` + tidy up
