+++ crates/polyfuse/src/nix.rs
@@ -1,6 +1,6 @@
 //! A small facilities of Unix syscalls.
 
-use libc::{c_int, c_void, iovec};
+use libc::{c_int, iovec};
 use std::{
     cmp, io,
     os::{fd::BorrowedFd, unix::prelude::*},
@@ -17,18 +17,6 @@ macro_rules! syscall {
     }};
 }
 
-#[allow(dead_code)]
-pub fn read(fd: BorrowedFd<'_>, buf: &mut [u8]) -> io::Result<usize> {
-    let len = syscall! {
-        read(
-            fd.as_raw_fd(), //
-            buf.as_mut_ptr() as *mut c_void,
-            buf.len(),
-        )
-    };
-    Ok(len as usize)
-}
-
 pub fn readv(fd: BorrowedFd<'_>, bufs: &mut [io::IoSliceMut<'_>]) -> io::Result<usize> {
     let len = syscall! {
         readv(
@@ -38,18 +26,6 @@ pub fn readv(fd: BorrowedFd<'_>, bufs: &mut [io::IoSliceMut<'_>]) -> io::Result<
         )
     };
     Ok(len as usize)
-}
-
-#[allow(dead_code)]
-pub fn write(fd: BorrowedFd<'_>, buf: &[u8]) -> io::Result<usize> {
-    let res = syscall! {
-        write(
-            fd.as_raw_fd(), //
-            buf.as_ptr() as *const c_void,
-            buf.len(),
-        )
-    };
-    Ok(res as usize)
 }
 
 pub fn writev(fd: BorrowedFd<'_>, bufs: &[io::IoSlice<'_>]) -> io::Result<usize> {
